Crate deaf

DEAF: The Definitive Elf Analysis Framework

Contribution Guide
Code of Conduct

This is a library for parsing and modifying ELF-format applications and libraries. There are many other libraries out there for parsing the ELF format, like the appropriately-named elf, elfy, and elfkit crates, but they are either not designed with modification in mind (elf/elfy) or are very limited (and now unmaintained) (elfkit). To fix this situation, I’m writing DEAF with the explicit and primary goal of making it possible to change anything you can see. This means that if you load a shared library and parse the dynamic symbol table, you should be able to-

Reorder

Remove

Replace

Update

Append

-symbols and then write the resulting symbol table back into a file. Existing toolkits generally don’t have this capability, or it’s very limited and oriented toward cybersecurity tasks like injecting telemetry or inserting malicious relocation records. If you want remove the body, symbol and relocation records of a particular function from a library, then shift all following functions forward, then update any relocation/init or preinit array references, you just can’t.

DEAF exists specifically to make things like this possible.

Notes

The name “DEAF” is a back-ronym- it doesn’t mean anything. It’s just a short word that reminded me of the DEADBEEF test value used in a lot of hex examples. A friend suggested “Definitive Elf Analysis Framework” as the meaning after I already had the name, and I couldn’t think of anything better.

Any time you see an attribute/method named kind, it’s a synonym for type (e.g. Relocation::kind) which is a reserved keyword in the rust language, but is used for a lot of fields in the ELF format.

References

Topic	Url

ELF Format	ELF_Format.pdf

Symbols	Oracle Docs: Symbols

Relocations	Oracle Docs: Relocations

Init Array	Oracle Docs: Init

Modules

arrays

Module that defines arrays (init, fini etc.) and array items

common

Common structs and types used by all other modules

errors

A custom error type that aggregates internal errors

headers

File, program and section header definitions

tables

Various types of tables and associated table items

utilities

Utility functions for both testing and execution

Structs

Binary

An ELF formatted binary file

Section

A Section extracted from an ELF file

Segment

A Segment extracted from an ELF file