Struct deaf::arrays::Array

pub struct Array { /* private fields */ }

A section interpreted as an array

Each Array instance is considered to be a series of ArrayItem structs. Arrays can be parsed from a SectionHeader and a byte buffer containing the body of the section.

Implementations

impl Array

pub fn new( offset: usize, size: usize, layout: Layout, width: Width, kind: SHType, entity_size: usize ) -> Self

Create a new array given type and position information

pub fn read(&mut self, bytes: &[u8]) -> Result<usize>

Read array items from a byte buffer

Reads from an offset to offset + section_size

pub fn write(&self, bytes: &mut [u8]) -> Result<usize>

Write array items to a byte buffer

Writes from the beginning of the given byte slice

pub fn len(&self) -> usize

The number of items in this array

pub fn size(&self) -> usize

The calculated size of the array in bytes

pub fn get(&self, index: usize) -> Option<&ArrayItem>

Get a reference to an array item at index

pub fn get_mut(&mut self, index: usize) -> Option<&mut ArrayItem>

Get a mutable reference to an array item at index

pub fn insert(&mut self, index: usize, item: ArrayItem)

Insert an array item at index, pushing other items right

pub fn push(&mut self, item: ArrayItem)

Append an array item to the array

pub fn remove(&mut self, index: usize) -> ArrayItem

Remove and return an array item from this array