deaf
Module common

Common structs and types used by all other modules

Re-exports

pub use ranges::Ranges;

Modules

ranges

Structs

All

Placeholder target for Update

ByteIter

An iterator that uses a delimiting value or length to convert a byte buffer into a series of slices.

Field

A single field in a section, table item etc.

Item

An item in a section, table item etc that contains a field and associated value.

ItemArray

Update

Update for an object of type T

Enums

ByteDelimiter

The delimiter used to determine the ending index of sub-slices returned by ByteIter.

Layout

The endianness of the ELF file

PHType

The type of a program header

SHFlags

The flags of a particular section header

SHIndex

Reserved values for section header indices

SHType

The type of a section header

STBind

The binding of a symbol entry from a static or dynamic symbol table

STType

The type of a symbol entry from a static or dynamic symbol table

STVisibility

The binding of a symbol entry from a static or dynamic symbol table

SectionType

Width

The mode of the ELF file

Traits

Convert

Convert a value into another value

FromBytes

Parse an object from bytes

IntoBytes

Convert an object into bytes

T32Value

T64Value

TOutValue

Updateable

Trait to add an update function to structs that can be updated.