Enum deaf::common::STType

#[repr(u8)]
pub enum STType {
    STT_NOTYPE,
    STT_OBJECT,
    STT_FUNC,
    STT_SECTION,
    STT_FILE,
    STT_COMMON,
    STT_TLS,
}

The type of a symbol entry from a static or dynamic symbol table

This enum is parsed from the lower 4 bits of the 1-byte ‘st_info’ field include in each entry of the symbol table.

Variants

STT_NOTYPE

STT_OBJECT

STT_FUNC

STT_SECTION

STT_FILE

STT_COMMON

STT_TLS