deaf::common
Module ranges

Structs

Ranges

This struct maintains several ranges and returns them depending on the current width.

Constants

ADDRESS

Range for extracting addresses from init/preinit/fini tables

EI_ABIVERSION

Further identifies specific version of ABI

EI_CLASS

The format of the file (1 = 32-bit, 2 = 64-bit)

EI_DATA

Endianness of the file (1 = little, 2 = big)

EI_KEY

Leading byte of the ELF “magic number”, should always be 0x7F

EI_MAGIC

The ASCII representation of the letters ELF

EI_OSABI

Identifies the target operating system ABI

EI_PAD

Padding bytes that should always be filled with zeroes

EI_VERSION

Version of ELF format (always 1)

E_EHSIZE

Size of the header

E_ENTRY

Offset of the entry point for execution (0 if none)

E_FLAGS

Target architecture-specific flags

E_MACHINE

Specifies target instruction set architecture

E_PHENTSIZE

Size of an entry in the program header

E_PHNUM

Number of items in the program header

E_PHOFF

Offset of the program header within the file

E_SHENTSIZE

Size of an entry in the section header

E_SHNUM

Number of items in the section header

E_SHOFF

Offset of the section header table

E_SHSTRNDX

Index of the section header that contains section names

E_TYPE

The type of the object file

E_VERSION

Set to 1 for the original ELF format

P_ALIGN

Alignment of the file in memory (0/1 = no alignment, otherwise positive power of 2)

P_FILESZ

Size in bytes of the segment in the file

P_FLAGS

Segment-specific flags

P_MEMSZ

Size in bytes of the segment in memory

P_OFFSET

Offset of the segment in the file

P_PADDR

Physical address of the segment

P_TYPE

Type of the segment

P_VADDR

Virtual address of the segment in memory

RT_ADDEND

Addend of the relocation, if found

RT_INFO

Relocation info for relocation entry

RT_OFFSET

Offset of the relocation entry

SH_ADDR

Virtual address of the section in memory

SH_ADDRALIGN

Required alignment of the section (must be power of 2)

SH_ENTSIZE

Contains size of entries in bytes, if the section has entries

SH_FLAGS

Attributes of the section header

SH_INFO

Extra information about the section

SH_LINK

Section index of an associated section

SH_NAME

Offset to a string name in the string table

SH_OFFSET

Offset of the section in the file image

SH_SIZE

The size of the section

SH_TYPE

The type of the section header

ST_INFO

Type and binding attributes of the symbol

ST_NAME

Index of the symbols name in the string table

ST_OTHER

Visibility of the symbol

ST_SHNDX

Index of the symbol’s section in the section header table

ST_SIZE

The size in bytes of the symbol data

ST_VALUE

The value of the symbol table entry