deaf::common
Struct Item

pub struct Item<T32 = u8, T64 = T32, Out = T64>
where
    T32: T32Value<Out>,
    T64: T64Value<Out>,
    Out: TOutValue<T32, T64>,
{ /* private fields */ }

An item in a section, table item etc that contains a field and associated value.

Implementations

impl<T32, T64, Out> Item<T32, T64, Out>
where T32: T32Value<Out>, T64: T64Value<Out>, Out: TOutValue<T32, T64>,

pub fn new(ranges: Ranges) -> Self

Create a new item with given ranges

pub fn make(ranges: Ranges, width: Width, layout: Layout) -> Self

Create a new item with ranges, width and layout

pub fn with_width(self, width: Width) -> Self

Builder method to set the initial width

pub fn with_layout(self, layout: Layout) -> Self

Builder method to set the initial layout

pub fn with_value(self, value: Out) -> Self

Builder method to set the initial value

pub fn with_offset(self, offset: usize) -> Self

Builder method to set an offset

pub fn with_index(self, index: usize) -> Self

Builder method to set an index

pub fn parse(self, bytes: &[u8]) -> Result<Self>

Builder method to parse a byte buffer

pub fn read(&mut self, bytes: &[u8]) -> Result<Out>

Read the value if possible

pub fn write(&self, bytes: &mut [u8]) -> Result<()>

Write the value if there is one

pub fn get(&self) -> Out

Get the output value of the item

pub fn set(&mut self, value: Out)

Set the output value of the item

pub fn size(&self) -> usize

Get the size of the item if there is a value

pub fn width(&self) -> Width

Get the width (32- or 64-bit) of the item

pub fn set_width(&mut self, width: Width)

Set the width (32- or 64-bit) of the item

pub fn layout(&self) -> Layout

Get the layout (little- or big-endian) of the item

pub fn set_layout(&mut self, layout: Layout)

Set the layout (little- or big-endian) of the item

pub fn set_offset(&mut self, offset: usize)

Set an offset in bytes to read at

pub fn set_index(&mut self, index: usize)

Set an index based on field size to read at