deaf::common
Struct ItemArray

pub struct ItemArray<T32, T64 = T32, Out = T64>
where
    T32: T32Value<Out>,
    T64: T64Value<Out>,
    Out: TOutValue<T32, T64>,
{ /* private fields */ }

An array of items in a section

Implementations

impl<T32, T64, Out> ItemArray<T32, T64, Out>
where T32: T32Value<Out>, T64: T64Value<Out>, Out: TOutValue<T32, T64>,

pub fn new(ranges: Ranges) -> Self

Create a new item array with given ranges

pub fn make(ranges: Ranges, width: Width, layout: Layout) -> Self

Create a new item array with given ranges, layout, and width

pub fn with_width(self, width: Width) -> Self

Builder method to set the initial width

pub fn with_layout(self, layout: Layout) -> Self

Builder method to set the initial layout

pub fn with_offset(self, offset: usize) -> Self

Builder method to add an array offset

pub fn with_last_offset(self, offset: usize) -> Self

Builder method to set the last offset for the array

pub fn iterator<'a>(&'a self, bytes: &'a [u8]) -> ByteIter<'a>

Get an iterator over slices of the data

pub fn read(&self, bytes: &[u8], index: usize) -> Result<Out>

Read a value at a particular index in the data

pub fn item_offset(&self, index: usize) -> usize

Get the offset of a particular index in the array

pub fn item_size(&self) -> usize

Get the size of the item for the current width

pub fn length(&self, count: usize) -> usize

Get the length of the array, given the item count

pub fn width(&self) -> Width

Get the width (32- or 64-bit) of the item

pub fn set_width(&mut self, width: Width)

Set the width (32- or 64-bit) of the item

pub fn layout(&self) -> Layout

Get the layout (little- or big-endian) of the item

pub fn set_layout(&mut self, layout: Layout)

Set the layout (little- or big-endian) of the item

pub fn offset(&self) -> usize

Get the total offset at which the array will read

pub fn add_offset(&mut self, offset: usize)

Set the offset of the array in the parsed data