Module deaf::headers::common::constants

Enums

Layout

The endianness of the ELF file

PHType

The type of a program header

SHFlags

The flags of a particular section header

SHType

The type of a section header

STBind

The binding of a symbol entry from a static or dynamic symbol table

STType

The type of a symbol entry from a static or dynamic symbol table

Width

The mode of the ELF file

Constants

FH_SIZE_32

FH_SIZE_64

PH_SIZE_32

PH_SIZE_64

RT_SIZE_32

RT_SIZE_64

SH_SIZE_32

SH_SIZE_64

ST_SIZE_32

ST_SIZE_64