Struct deaf::headers::FileHeader

pub struct FileHeader { /* private fields */ }

The ELF file header parsed from the beginning of the file

Implementations

impl FileHeader

pub fn new() -> Self

Create a new header with given Layout and Width

All fields are None until read

pub fn parse(b: &[u8]) -> Result<Self>

Parse a header from the provided byte buffer

pub fn read(&mut self, b: &[u8]) -> Result<()>

Read values from a byte buffer

Byte buffer is assumed to be sliced such that the header is at the beginning of the buffer.

pub fn write(&mut self, b: &mut [u8]) -> Result<()>

Write values to a byte buffer

Byte buffer is assumed to be sliced such that the header will be written at the correct position.

pub fn size(&self) -> usize

The size of the header in bytes

pub fn width(&self) -> Width

Get the width (32 or 64-bit) of the header

pub fn set_width(&mut self, width: Width)

Set the width of the header

pub fn layout(&self) -> Layout

Get the layout (little or big-endian) of the header

pub fn set_layout(&mut self, layout: Layout)

Set the layout of the header

pub fn magic(&self) -> Option<String>

Get the ei_magic attribute of the header

pub fn set_magic(&mut self, magic: String)

Set the ei_magic attribute of the header

pub fn class(&self) -> Option<Width>

Get the ei_class attribute of the header

pub fn set_class(&mut self, class: Width)

Set the ei_class attribute of the header

pub fn data(&self) -> Option<Layout>

Get the ei_data attribute of the header

pub fn set_data(&mut self, data: Layout)

Set the ei_data attribute of the header

pub fn version(&self) -> Option<u8>

Get the ei_version attribute of the header

pub fn set_version(&mut self, version: u8)

Set the ei_version attribute of the header

pub fn osabi(&self) -> Option<u8>

Get the ei_osabi attribute of the header

pub fn set_osabi(&mut self, osabi: u8)

Set the ei_osabi attribute of the header

pub fn abiversion(&self) -> Option<u8>

Get the abiversion attribute of the header

pub fn set_abiversion(&mut self, abiversion: u8)

Set the abiversion attribute of the header

pub fn file_type(&self) -> Option<u16>

Get the e_type attribute of the header

pub fn set_file_type(&mut self, file_type: u16)

Set the e_type attribute of the header

pub fn machine(&self) -> Option<u16>

Get the e_machine attribute of the header

pub fn set_machine(&mut self, machine: u16)

Set the e_machine attribute of the header

pub fn entry(&self) -> Option<u64>

Get the e_entry attribute of the header

pub fn set_entry(&mut self, entry: u64)

Set the e_entry attribute of the header

pub fn phoff(&self) -> Option<usize>

Get the e_phoff attribute of the header

pub fn set_phoff(&mut self, phoff: usize)

Set the e_phoff attribute of the header

pub fn shoff(&self) -> Option<usize>

Get the e_shoff attribute of the header

pub fn set_shoff(&mut self, shoff: usize)

Set the e_shoff attribute of the header

pub fn flags(&self) -> Option<u32>

Get the e_flags attribute of the header

pub fn set_flags(&mut self, flags: u32)

Set the e_flags attribute of the header

pub fn ehsize(&self) -> Option<u16>

Get the e_ehsize attribute of the header

pub fn set_ehsize(&mut self, ehsize: u16)

Set the e_ehsize attribute of the header

pub fn phentsize(&self) -> Option<usize>

Get the e_phentsize attribute of the header

pub fn set_phentsize(&mut self, phentsize: usize)

Set the e_phentsize attribute of the header

pub fn phnum(&self) -> Option<usize>

Get the e_phnum attribute of the header

pub fn set_phnum(&mut self, phnum: usize)

Set the e_phnum attribute of the header

pub fn shentsize(&self) -> Option<usize>

Get the e_shentsize attribute of the header

pub fn set_shentsize(&mut self, shentsize: usize)

Set the e_shentsize attribute of the header

pub fn shnum(&self) -> Option<usize>

Get the e_shnum attribute of the header

pub fn set_shnum(&mut self, shnum: usize)

Set the e_shnum attribute of the header

pub fn shstrndx(&self) -> Option<usize>

Get the e_shstrndx attribute of the header

pub fn set_shstrndx(&mut self, shstrndx: usize)

Set the e_shstrndx attribute of the header