Struct deaf::headers::SectionHeader

pub struct SectionHeader { /* private fields */ }

Section headers extracted from an ELF file.

Normally found at the offset declared in the FileHeader as ‘shoff’.

Implementations

impl SectionHeader

pub fn new(layout: Layout, width: Width) -> Self

Create a new header with given Layout and Width

All fields are None until read

pub fn parse(b: &[u8], layout: Layout, width: Width) -> Result<Self>

Parse a header from the provided byte buffer

pub fn parse_all( b: &[u8], count: usize, offset: usize, size: usize, layout: Layout, width: Width ) -> Result<Vec<Self>>

Parse all headers for a byte array given count, offset etc.

pub fn read(&mut self, b: &[u8]) -> Result<()>

Read values from a byte buffer

Byte buffer is assumed to be sliced such that the header is at the beginning of the buffer.

pub fn write(&self, b: &mut [u8]) -> Result<()>

Write values to a byte buffer

Byte buffer is assumed to be sliced such that the header will be written at the correct position.

pub fn size(&self) -> usize

The size of the header in bytes

pub fn width(&self) -> Width

Get the width (32 or 64-bit) of the header

pub fn set_width(&mut self, width: Width)

Set the width of the header

pub fn layout(&self) -> Layout

Get the layout (little or big-endian) of the header

pub fn set_layout(&mut self, layout: Layout)

Set the layout of the header

pub fn name(&self) -> u32

Get the sh_name attribute of the header

pub fn set_name(&mut self, name: u32)

Set the sh_name attribute of the header

pub fn kind(&self) -> SHType

Get the sh_type attribute of the header

pub fn set_kind(&mut self, kind: SHType)

Set the sh_type attribute of the header

pub fn flags(&self) -> BitFlags<SHFlags>

Get the sh_flags attribute of the header

pub fn set_flags(&mut self, flags: BitFlags<SHFlags>)

Set the sh_flags attribute of the header

pub fn address(&self) -> u64

Get the sh_address attribute of the header

pub fn set_address(&mut self, address: u64)

Set the sh_address attribute of the header

pub fn offset(&self) -> usize

Get the sh_offset attribute of the header

pub fn set_offset(&mut self, offset: usize)

Set the sh_offset attribute of the header

pub fn body_size(&self) -> usize

Get the sh_size attribute of the header

pub fn set_body_size(&mut self, body_size: usize)

Set the sh_size attribute of the header

pub fn link(&self) -> u32

Get the sh_link attribute of the header

pub fn set_link(&mut self, link: u32)

Set the sh_link attribute of the header

pub fn info(&self) -> u32

Get the sh_info attribute of the header

pub fn set_info(&mut self, info: u32)

Set the sh_info attribute of the header

pub fn addralign(&self) -> u64

Get the sh_addralign attribute of the header

pub fn set_addralign(&mut self, addralign: u64)

Set the sh_addralign attribute of the header

pub fn entsize(&self) -> usize

Get the sh_entsize attribute of the header

pub fn set_entsize(&mut self, entsize: usize)

Set the sh_entsize attribute of the header