deaf
Struct Binary

pub struct Binary { /* private fields */ }

An ELF formatted binary file

Implementations

impl Binary

pub fn read(&mut self, data: &[u8]) -> Result<usize>

pub fn write(&self, data: &mut [u8]) -> Result<usize>

pub fn load<T: AsRef<Path>>(path: T) -> Result<Self>

pub fn save<T: AsRef<Path>>(&self, path: T) -> Result<usize>

pub fn process(&mut self) -> Result<()>

pub fn size(&self) -> usize

pub fn section(&self, index: usize) -> Result<&Section>

pub fn section_for_address(&self, address: usize) -> Result<&Section>

pub fn section_mut(&mut self, index: usize) -> Result<&mut Section>

pub fn sections(&self, kind: SectionType) -> Vec<&Section>

pub fn sections_mut(&mut self, kind: SectionType) -> Vec<&mut Section>

pub fn section_name(&self, offset: usize) -> Result<String>

pub fn section_names(&self) -> Result<Vec<String>>

pub fn string_tables(&self) -> Vec<StringTable<'_>>

Get all string tables except the ‘shstrtab’

pub fn symbol_tables(&self) -> Vec<SymbolTable<'_>>

pub fn symbol_tables_mut(&mut self) -> Vec<SymbolTableMut<'_>>

pub fn symbol_name(&self, offset: usize) -> Result<String>

pub fn symbols(&self) -> Result<Vec<Symbol>>

pub fn functions(&self) -> Result<Vec<Function>>

pub fn data(&self, address: usize, size: usize) -> Vec<u8>

pub fn shnum(&self) -> usize

Get the number of section headers in the file

pub fn shoff(&self) -> usize

Get the offset of the section header table

pub fn shentsize(&self) -> usize

Get the size of section headers

pub fn phnum(&self) -> usize

Get the number of program headers in the file

pub fn phoff(&self) -> usize

Get the offset of the program header table

pub fn phentsize(&self) -> usize

Get the size of program headers

pub fn shstrndx(&self) -> usize

pub fn layout(&self) -> Layout

Get the layout of the file (little or big endian)

pub fn width(&self) -> Width

Get the addressing width of the file (32, 64 etc)