deaf
Module tables

Various types of tables and associated table items

Each table type can be created (using TryFrom) from an appropriate section struct. TryFrom will fail if the section isn’t the correct type (e.g. non-SHT_SYMTAB section header cannot be converted into a Table<SymbolItem> struct).

Structs

ArrayItem

A specialization of item for reading addresses from Arrays

RelItem

A Relocation item found in relocation tables

RelaItem

A Relocation item found in relocation tables

RelocationInfo

Representation of the info field in a Relocation record

StringItem

A String item found in string tables

Table

A Section represented as an immutable Table

TableMut

A Section represented as a mutable Table

Traits

TableItem

A record that can be extracted from a table section

TableView

Shared table interface between Table and TableMut

Type Aliases

Array

ArrayMut

RelTable

RelTableMut

RelaTable

RelaTableMut

StringTable

StringTableMut

SymbolTable

SymbolTableMut