Trait deaf::tables::common::Table

pub trait Table<T> {
    fn len(&self) -> usize;
    fn size(&self) -> usize;
    fn get(&self, index: usize) -> Option<T>;
    fn set(&mut self, index: usize, item: T) -> Result<usize>;
    fn add(&mut self, item: T) -> Result<usize>;
    fn del(&mut self, index: usize) -> Option<T>;
}

Required Methods

fn len(&self) -> usize

the number of elements in the table

fn size(&self) -> usize

the size in bytes of the table

fn get(&self, index: usize) -> Option<T>

get the item at index or return None

fn set(&mut self, index: usize, item: T) -> Result<usize>

set an index position to item returning the index on success

fn add(&mut self, item: T) -> Result<usize>

add the item at the end of the table returning the index on success

fn del(&mut self, index: usize) -> Option<T>

Delete an element from the table, returning the element on success or None