Module deaf::tables

Various types of tables and associated table items

Each table type can be created (using TryFrom) from an appropriate section header struct. TryFrom will fail if the section isn’t the correct type (e.g. non-SHT_SYMTAB section header cannot be converted into a Table<SymbolItem> struct).

Structs

RelItem

A Relocation item found in relocation tables

RelaItem

A Relocation item found in relocation tables

RelocationInfo

Representation of the info field in a Relocation record

StringItem

A String item found in string tables

SymbolInfo

Representation of the info field in a Symbol record

SymbolItem

A Symbol item found in symbol tables

Table

A section interpreted as a table

Traits

TableItem

A record that can be extracted from a table section

Type Definitions

RelTable

Alias for a Table that contains RelItem records

RelaTable

Alias for a Table that contains RelaItem records

StringTable

Alias for a Table that contains StringItem records

SymbolTable

Alias for a Table that contains SymbolItem records