deaf::tables
Trait TableView

pub trait TableView<T>
where
    T: TableItem + Default,
{

    // Required method
    fn section(&self) -> &Section;

    // Provided methods
    fn name_index(&self) -> usize { ... }
    fn iterator(&self) -> ByteIter<'_>  { ... }
    fn data(&self, index: usize) -> Result<&[u8]> { ... }
    fn offset(&self, index: usize) -> usize { ... }
    fn at(&self, index: usize) -> Result<T> { ... }
    fn at_offset(&self, offset: usize) -> Result<T> { ... }
    fn items(&self) -> Result<Vec<T>> { ... }
    fn len(&self) -> usize { ... }
    fn size(&self) -> usize { ... }
    fn layout(&self) -> Layout { ... }
    fn width(&self) -> Width { ... }
    fn has_fixed_size(&self) -> bool { ... }
    fn has_variable_size(&self) -> bool { ... }

}

Shared table interface between Table and TableMut

Required Methods

fn section(&self) -> &Section

Get an immutable reference to the internal section

Provided Methods

fn name_index(&self) -> usize

Get the name index of the internal section

fn iterator(&self) -> ByteIter<'_>

Get an iterator over each item’s binary data

fn data(&self, index: usize) -> Result<&[u8]>

Get a slice of data that represents an item

fn offset(&self, index: usize) -> usize

Get the offset of an item from the index

fn at(&self, index: usize) -> Result<T>

Get an element from the table

fn at_offset(&self, offset: usize) -> Result<T>

Get an element from the table at a byte offset

fn items(&self) -> Result<Vec<T>>

Get all items from the table

fn len(&self) -> usize

Get the number of items in the table

fn size(&self) -> usize

Get the number of bytes in the table

fn layout(&self) -> Layout

Get the layout being used by this table

fn width(&self) -> Width

Get the width being used by this table

fn has_fixed_size(&self) -> bool

True if items are all the same size

fn has_variable_size(&self) -> bool

True if items can be different sizes